requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
executing==0.8.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
traitlets==5.3.0
tzdata==2022.1
wcwidth==0.2.5
